refactor: introduce engine handler (#6680)
* refactor: introduce engine handler


* clean


---------
