+++ pkg/engine/mutation_test.go
@@ -113,11 +113,11 @@ func Test_VariableSubstitutionPatchStrategicMerge(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	policyContext := &PolicyContext{
-		policy:      &policy,
-		jsonContext: ctx,
-		newResource: *resourceUnstructured,
-	}
+
+	policyContext := NewPolicyContextWithJsonContext(ctx).
+		WithPolicy(&policy).
+		WithNewResource(*resourceUnstructured)
+
 	er := testMutate(context.TODO(), nil, registryclient.NewOrDie(), policyContext, nil)
 	t.Log(string(expectedPatch))
 
@@ -187,11 +187,10 @@ func Test_variableSubstitutionPathNotExist(t *testing.T) {
 	err = enginecontext.AddResource(ctx, resourceRaw)
 	assert.NilError(t, err)
 
-	policyContext := &PolicyContext{
-		policy:      &policy,
-		jsonContext: ctx,
-		newResource: *resourceUnstructured,
-	}
+	policyContext := NewPolicyContextWithJsonContext(ctx).
+		WithPolicy(&policy).
+		WithNewResource(*resourceUnstructured)
+
 	er := testMutate(context.TODO(), nil, registryclient.NewOrDie(), policyContext, nil)
 	assert.Equal(t, len(er.PolicyResponse.Rules), 1)
 	assert.Assert(t, strings.Contains(er.PolicyResponse.Rules[0].Message, "Unknown key \"name1\" in path"))
@@ -265,11 +264,9 @@ func Test_variableSubstitutionCLI(t *testing.T) {
 	err = enginecontext.AddResource(ctx, resourceRaw)
 	assert.NilError(t, err)
 
-	policyContext := &PolicyContext{
-		policy:      &policy,
-		jsonContext: ctx,
-		newResource: *resourceUnstructured,
-	}
+	policyContext := NewPolicyContextWithJsonContext(ctx).
+		WithPolicy(&policy).
+		WithNewResource(*resourceUnstructured)
 
 	er := testMutate(
 		context.TODO(),
@@ -387,11 +384,9 @@ func Test_chained_rules(t *testing.T) {
 	err = ctx.AddResource(resource.Object)
 	assert.NilError(t, err)
 
-	policyContext := &PolicyContext{
-		policy:      &policy,
-		jsonContext: ctx,
-		newResource: *resource,
-	}
+	policyContext := NewPolicyContextWithJsonContext(ctx).
+		WithPolicy(&policy).
+		WithNewResource(*resource)
 
 	err = ctx.AddImageInfos(resource, cfg)
 	assert.NilError(t, err)
@@ -477,11 +472,9 @@ func Test_precondition(t *testing.T) {
 	err = enginecontext.AddResource(ctx, resourceRaw)
 	assert.NilError(t, err)
 
-	policyContext := &PolicyContext{
-		policy:      &policy,
-		jsonContext: ctx,
-		newResource: *resourceUnstructured,
-	}
+	policyContext := NewPolicyContextWithJsonContext(ctx).
+		WithPolicy(&policy).
+		WithNewResource(*resourceUnstructured)
 
 	er := testMutate(context.TODO(), nil, registryclient.NewOrDie(), policyContext, enginetest.ContextLoaderFactory(nil, nil))
 	t.Log(string(expectedPatch))
@@ -573,11 +566,9 @@ func Test_nonZeroIndexNumberPatchesJson6902(t *testing.T) {
 	err = enginecontext.AddResource(ctx, resourceRaw)
 	assert.NilError(t, err)
 
-	policyContext := &PolicyContext{
-		policy:      &policy,
-		jsonContext: ctx,
-		newResource: *resourceUnstructured,
-	}
+	policyContext := NewPolicyContextWithJsonContext(ctx).
+		WithPolicy(&policy).
+		WithNewResource(*resourceUnstructured)
 
 	er := testMutate(context.TODO(), nil, registryclient.NewOrDie(), policyContext, enginetest.ContextLoaderFactory(nil, nil))
 	t.Log(string(expectedPatch))
@@ -661,11 +652,9 @@ func Test_foreach(t *testing.T) {
 	err = ctx.AddResource(resource.Object)
 	assert.NilError(t, err)
 
-	policyContext := &PolicyContext{
-		policy:      &policy,
-		jsonContext: ctx,
-		newResource: *resource,
-	}
+	policyContext := NewPolicyContextWithJsonContext(ctx).
+		WithPolicy(&policy).
+		WithNewResource(*resource)
 
 	err = ctx.AddImageInfos(resource, cfg)
 	assert.NilError(t, err)
@@ -765,11 +754,9 @@ func Test_foreach_element_mutation(t *testing.T) {
 	err = ctx.AddResource(resource.Object)
 	assert.NilError(t, err)
 
-	policyContext := &PolicyContext{
-		policy:      &policy,
-		jsonContext: ctx,
-		newResource: *resource,
-	}
+	policyContext := NewPolicyContextWithJsonContext(ctx).
+		WithPolicy(&policy).
+		WithNewResource(*resource)
 
 	err = ctx.AddImageInfos(resource, cfg)
 	assert.NilError(t, err)
@@ -888,11 +875,9 @@ func Test_Container_InitContainer_foreach(t *testing.T) {
 	err = ctx.AddResource(resource.Object)
 	assert.NilError(t, err)
 
-	policyContext := &PolicyContext{
-		policy:      &policy,
-		jsonContext: ctx,
-		newResource: *resource,
-	}
+	policyContext := NewPolicyContextWithJsonContext(ctx).
+		WithPolicy(&policy).
+		WithNewResource(*resource)
 
 	err = ctx.AddImageInfos(resource, cfg)
 	assert.NilError(t, err)
@@ -1035,11 +1020,9 @@ func testApplyPolicyToResource(t *testing.T, policyRaw, resourceRaw []byte) engi
 	err = ctx.AddResource(resource.Object)
 	assert.NilError(t, err)
 
-	policyContext := &PolicyContext{
-		policy:      &policy,
-		jsonContext: ctx,
-		newResource: *resource,
-	}
+	policyContext := NewPolicyContextWithJsonContext(ctx).
+		WithPolicy(&policy).
+		WithNewResource(*resource)
 
 	err = ctx.AddImageInfos(resource, cfg)
 	assert.NilError(t, err)
@@ -1588,11 +1571,9 @@ func Test_mutate_existing_resources(t *testing.T) {
 			_, err = dclient.GetResource(context.TODO(), target.GetAPIVersion(), target.GetKind(), target.GetNamespace(), target.GetName())
 			assert.NilError(t, err)
 
-			policyContext = &PolicyContext{
-				policy:      &policy,
-				jsonContext: ctx,
-				newResource: *trigger,
-			}
+			policyContext = NewPolicyContextWithJsonContext(ctx).
+				WithPolicy(&policy).
+				WithNewResource(*trigger)
 
 			er := testMutate(context.TODO(), dclient, registryclient.NewOrDie(), policyContext, nil)
 
@@ -1697,11 +1678,9 @@ func Test_RuleSelectorMutate(t *testing.T) {
 	_, err = ctx.Query("request.object.metadata.name")
 	assert.NilError(t, err)
 
-	policyContext := &PolicyContext{
-		policy:      &policy,
-		jsonContext: ctx,
-		newResource: *resourceUnstructured,
-	}
+	policyContext := NewPolicyContextWithJsonContext(ctx).
+		WithPolicy(&policy).
+		WithNewResource(*resourceUnstructured)
 
 	er := testMutate(context.TODO(), nil, registryclient.NewOrDie(), policyContext, nil)
 	assert.Equal(t, len(er.PolicyResponse.Rules), 2)
@@ -1716,7 +1695,7 @@ func Test_RuleSelectorMutate(t *testing.T) {
 	}
 
 	applyOne := kyverno.ApplyOne
-	policyContext.policy.GetSpec().ApplyRules = &applyOne
+	policyContext.Policy().GetSpec().ApplyRules = &applyOne
 
 	er = testMutate(context.TODO(), nil, registryclient.NewOrDie(), policyContext, nil)
 	assert.Equal(t, len(er.PolicyResponse.Rules), 1)
@@ -2078,11 +2057,9 @@ func Test_SpecialCharacters(t *testing.T) {
 			}
 
 			// Create policy context.
-			policyContext := &PolicyContext{
-				policy:      &policy,
-				jsonContext: ctx,
-				newResource: *resource,
-			}
+			policyContext := NewPolicyContextWithJsonContext(ctx).
+				WithPolicy(&policy).
+				WithNewResource(*resource)
 
 			// Mutate and make sure that we got the expected amount of rules.
 			patches := testMutate(context.TODO(), nil, registryclient.NewOrDie(), policyContext, nil).GetPatches()
