+++ pkg/engine/resources/resources.go
@@ -1,0 +1,8 @@
+package resources
+
+import (
+	_ "embed"
+)
+
+//go:embed default-config.yaml
+var DefaultConfigBytes []byte
